# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "YouTube Downloader API",
    "description": "A comprehensive YouTube video and playlist downloader API for RapidAPI",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/youtube-downloader-api"
    }
  },
  "servers": [
    {
      "url": "https://your-api-url.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API Information",
        "description": "Get API information and available endpoints",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "endpoints": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Check API health status",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/video/info": {
      "get": {
        "summary": "Get Video Information",
        "description": "Get detailed information about a YouTube video",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "YouTube video URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_formats",
            "in": "query",
            "required": false,
            "description": "Include available download formats",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "integer"
                    },
                    "view_count": {
                      "type": "integer"
                    },
                    "like_count": {
                      "type": "integer"
                    },
                    "uploader": {
                      "type": "string"
                    },
                    "upload_date": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "age_limit": {
                      "type": "integer"
                    },
                    "webpage_url": {
                      "type": "string"
                    },
                    "formats": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid URL"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/video/download": {
      "get": {
        "summary": "Get Video Download Links",
        "description": "Get download links for a YouTube video",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "YouTube video URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "description": "Video quality",
            "schema": {
              "type": "string",
              "enum": ["highest", "high", "medium", "low", "audio_only"],
              "default": "high"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Preferred format",
            "schema": {
              "type": "string",
              "enum": ["mp4", "webm", "mkv", "mp3"],
              "default": "mp4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "integer"
                    },
                    "filesize": {
                      "type": "integer"
                    },
                    "ext": {
                      "type": "string"
                    },
                    "format_id": {
                      "type": "string"
                    },
                    "quality": {
                      "type": "string"
                    },
                    "requested_format": {
                      "type": "string"
                    },
                    "download_url": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid URL or quality"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/playlist/info": {
      "get": {
        "summary": "Get Playlist Information",
        "description": "Get information about a YouTube playlist",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "YouTube playlist URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of videos to retrieve",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "uploader": {
                      "type": "string"
                    },
                    "video_count": {
                      "type": "integer"
                    },
                    "videos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "duration": {
                            "type": "integer"
                          },
                          "uploader": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid playlist URL"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/playlist/download": {
      "get": {
        "summary": "Get Playlist Download Links",
        "description": "Get download links for videos in a YouTube playlist",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "YouTube playlist URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "description": "Video quality",
            "schema": {
              "type": "string",
              "enum": ["highest", "high", "medium", "low", "audio_only"],
              "default": "high"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Preferred format",
            "schema": {
              "type": "string",
              "enum": ["mp4", "webm", "mkv", "mp3"],
              "default": "mp4"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of videos to process",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist_info": {
                      "type": "object"
                    },
                    "download_links": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total_processed": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid playlist URL or parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          }
        }
      }
    }
  }
}